pyproject.toml
[project]
name = "eirt"
version = "0.1.0"
description = "EIRT PointPainting Implementation for KITTI Dataset"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bagpy>=0.5",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "open3d>=0.17.0",
    "tqdm>=4.65.0",
]
